Hygiene: remove unused classes and move card classes to `feed/model`
