+++ app/src/main/java/org/wikipedia/dataclient/RestService.kt
@@ -10,8 +10,6 @@ import org.wikipedia.dataclient.restbase.PreviewRequest
 import org.wikipedia.dataclient.restbase.RbDefinition
 import org.wikipedia.dataclient.restbase.UserEdits
 import org.wikipedia.feed.aggregated.AggregatedFeedContent
-import org.wikipedia.feed.announcement.AnnouncementList
-import org.wikipedia.feed.configure.FeedAvailability
 import org.wikipedia.feed.onthisday.OnThisDay
 import org.wikipedia.gallery.MediaList
 import org.wikipedia.readinglist.sync.SyncedReadingLists
@@ -84,11 +82,6 @@ interface RestService {
     suspend fun getOnThisDay(@Path("mm") month: Int,
                              @Path("dd") day: Int): OnThisDay
 
-    // TODO: Remove this before next fundraising campaign in 2024
-    @GET("feed/announcements")
-    @Headers("Accept: " + ACCEPT_HEADER_PREFIX + "announcements/0.1.0\"")
-    suspend fun getAnnouncements(): AnnouncementList
-
     @Headers("Accept: " + ACCEPT_HEADER_PREFIX + "aggregated-feed/0.5.0\"")
     @GET("feed/featured/{year}/{month}/{day}")
     suspend fun getFeedFeatured(
@@ -97,9 +90,6 @@ interface RestService {
         @Path("day") day: String?,
         @Query("lang") lang: String?
     ): AggregatedFeedContent
-
-    @GET("feed/availability")
-    suspend fun feedAvailability(): FeedAvailability
 
     // ------- Reading lists -------
     @POST("data/lists/setup")
