--- app/src/main/java/org/wikipedia/feed/announcement/AnnouncementList.kt
@@ -1,11 +1,0 @@
-package org.wikipedia.feed.announcement
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-class AnnouncementList {
-
-    @SerialName("announce")
-    val items: List<Announcement> = emptyList()
-}
