--- app/src/main/java/org/wikipedia/feed/dayheader/DayHeaderCard.kt
@@ -1,5 +1,0 @@
-package org.wikipedia.feed.dayheader
-
-import org.wikipedia.feed.model.Card
-
-class DayHeaderCard(val age: Int) : Card()
