+++ app/src/main/java/org/wikipedia/feed/model/DayHeaderCard.kt
@@ -1,0 +1,3 @@
+package org.wikipedia.feed.model
+
+class DayHeaderCard(val age: Int) : Card()
